Moving files to behave like the new folder structure
