+++ client/src/api/datasource/__tests__/ApiClient.test.tsx
@@ -1,6 +1,6 @@
 import { test, describe } from 'vitest';
 import nock from 'nock';
-import { ApiClient } from '../ApiClient';
+import { ApiClient } from '@/api/datasource/ApiClient';
 import { DataSource } from '@/api/Models';
 
 describe('ApiClient Metadata Tests', () => {
