+++ client/src/pages/Plugins.tsx
