+++ client/src/pages/signalGenerator/examples/README.md
