+++ client/src/pages/signalGenerator/examples/bpskToneNoise.ts
