+++ client/src/pages/signalGenerator/examples/exampleList.ts
